# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Booking Documentation",
        "description": "Hospital Booking API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://hospital-booking.local",
            "description": "local server"
        },
        {
            "url": "https://hospital.goharabbas.com/",
            "description": "production server"
        }
    ],
    "paths": {
        "/api/v1/hospitals": {
            "get": {
                "tags": [
                    "Hospitals"
                ],
                "summary": "Get all hospitals (public endpoint)",
                "description": "Returns a list of hospitals with their details (no authentication required)",
                "operationId": "a2f94899f5343cdfffd0965aa6b71098",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term to search across all text fields",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter by city name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "specialization",
                        "in": "query",
                        "description": "Filter by specialization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "facility",
                        "in": "query",
                        "description": "Filter by facility",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by hospital name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by (name_en, name_ar, city_en, city_ar)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "name_en"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Sort direction (asc, desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hospitals retrieved successfully"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/hospitals/{id}": {
            "get": {
                "tags": [
                    "Hospitals"
                ],
                "summary": "Get a specific hospital by ID (public endpoint)",
                "description": "Returns a single hospital with its details (no authentication required)",
                "operationId": "157781a7c6f017c6eab2a131b3560613",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of hospital to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hospital retrieved successfully"
                    },
                    "404": {
                        "description": "Hospital not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Hospitals",
            "description": "API endpoints for managing hospitals"
        }
    ]
}